pyproject.toml
[project]
name = "mdzen"
version = "0.2.0"
description = "MDZen - AI Agent for Molecular Dynamics Setup (MD + 膳/禅)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # MCP Server (FastMCP)
    # Note: fastmcp 2.x is required for Streamable HTTP transport (used in Jupyter/Colab)
    # The opentelemetry version conflict with google-adk is harmless (just warnings)
    "fastmcp>=2.0.0",
    "mcp[cli]>=1.18.0",

    # Google Agent Development Kit (ADK)
    "google-adk>=0.5.0",
    "litellm>=1.50.0,<2.0.0",

    # Constrain transitive dependencies to help pip resolution
    # These prevent "resolution-too-deep" errors
    "pyasn1>=0.6.0",
    "pyasn1-modules>=0.4.0",
    "google-auth>=2.30.0",
    "grpcio>=1.67.0,<1.68.0",  # litellm requires <1.68.0
    "grpcio-status>=1.67.0,<1.68.0",
    "opentelemetry-api==1.37.0",  # google-adk pins this version
    "opentelemetry-sdk==1.37.0",
    "opentelemetry-semantic-conventions==0.58b0",

    # Structure prediction & modeling
    # Note: Install openmm, rdkit, mdanalysis, biopython, pandas, numpy, scipy, pdbfixer via conda first
    # conda install -c conda-forge openmm rdkit mdanalysis biopython pandas numpy scipy openblas pdbfixer
    # Boltz-2: Install manually when needed (Phase 2-3)
    #   pip install 'boltz[cuda]' --no-deps
    #   Then install missing dependencies individually if needed
    "pdb2pqr>=3.1.0",
    "propka>=3.5.0",
    "gemmi>=0.6.0",  # mmCIF parsing for Boltz-2 output
    "dimorphite-dl>=1.3.2",  # pH-dependent protonation state generation
    "pubchempy>=1.0.4",  # PubChem compound search (used by genesis_server)
    "tavily-python>=0.3.0",  # Web search (used by research_server)

    # Topology manipulation
    "parmed>=4.3.0",

    # LLM integration
    "anthropic>=0.40.0",
    "prompt-toolkit>=3.0.0",

    # Data models & validation
    "pydantic>=2.12.3",

    # CLI & UI
    "typer>=0.19.2",
    "rich>=14.2.0",

    # Utilities
    "httpx>=0.28.1",
    "requests>=2.31.0",
    "pyyaml>=6.0",

    # Notebook development
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "python-dotenv>=1.0.0",

    # Visualization & analysis (for test notebooks)
    "py3Dmol>=2.0.0",
    "mdtraj>=1.9.0",
]

[project.optional-dependencies]
# Heavy dependencies (install via conda)
heavy = [
    "openmm>=8.3.1",
    "mdanalysis>=2.10.0",
    "rdkit>=2023.9.1",
    "biopython>=1.83",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    # ADK evaluation (optional, but required for `tests/adk_eval/**`)
    "google-adk[eval]>=0.5.0",
    "rouge-score>=0.1.2",
]

# Optional: install only evaluation dependencies (without full dev set)
eval = [
    "google-adk[eval]>=0.5.0",
    "rouge-score>=0.1.2",
]

[project.scripts]
mdzen = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["servers", "common", "src/mdzen"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
